Feature/front/camera 진짜로 merge중입니다..

See merge request s12-webmobile1-sub1/S12P11E106!44

+++ backend/yourcolors/src/main/java/com/ssafy/yourcolors/domain/info/entity/ProductColorId.java
@@ -10,6 +10,6 @@ import java.io.Serializable;
 @AllArgsConstructor
 @EqualsAndHashCode
 public class ProductColorId implements Serializable {
-    private Long product;
+    private int product;
     private String color;
 }
